refactoring to provide 4 different cse adapter services (idcc-import, idcc-export, d2cc-import and d2cc-export) based on config values
