+++ cse-adapter-app/src/main/java/com/farao_community/farao/cse/adapter/app/CseIdccExportService.java
@@ -5,10 +5,9 @@ import com.farao_community.farao.cse.runner.api.resource.CseExportResponse;
 import com.farao_community.farao.cse.runner.api.resource.ProcessType;
 import com.farao_community.farao.cse.runner.starter.CseClient;
 import com.farao_community.farao.gridcapa.task_manager.api.TaskDto;
-import org.apache.commons.lang3.NotImplementedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.stereotype.Service;
 
 import java.util.HashMap;
@@ -20,52 +19,34 @@ import java.util.concurrent.CompletableFuture;
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
 @Service
-@ConditionalOnBean(value = { CseExportAdapterConfiguration.class })
-public class CseExportService implements CseAdapter {
-    private static final Logger LOGGER = LoggerFactory.getLogger(CseExportService.class);
+@ConditionalOnExpression("'${cse-adapter.process-type}'.equalsIgnoreCase('IDCC') and '${cse-adapter.exchange-type}'.equalsIgnoreCase('export')")
+public class CseIdccExportService implements CseAdapter {
+    private static final Logger LOGGER = LoggerFactory.getLogger(CseIdccExportService.class);
 
-    private final CseExportAdapterConfiguration configuration;
     private final CseClient cseClient;
 
-    public CseExportService(CseExportAdapterConfiguration configuration, CseClient cseClient) {
-        this.configuration = configuration;
+    public CseIdccExportService(CseClient cseClient) {
         this.cseClient = cseClient;
     }
 
     @Override
     public void runAsync(TaskDto taskDto) {
-        CseExportRequest cseExportRequest;
-        switch (configuration.getProcessType()) {
-            case IDCC:
-                LOGGER.info("Sending export IDCC request for TS: {}", taskDto.getTimestamp());
-                cseExportRequest = getIdccRequest(taskDto);
-                break;
-            case D2CC:
-                LOGGER.info("Sending export D2CC request for TS: {}", taskDto.getTimestamp());
-                cseExportRequest = getD2ccRequest(taskDto);
-                break;
-            default:
-                throw new NotImplementedException(String.format("Unknown target process for CSE: %s", configuration.getProcessType()));
-        }
+        LOGGER.info("Sending export IDCC request for TS: {}", taskDto.getTimestamp());
+        CseExportRequest cseExportRequest = getIdccRequest(taskDto);
         CompletableFuture.runAsync(() -> cseClient.run(cseExportRequest, CseExportRequest.class, CseExportResponse.class));
     }
 
     CseExportRequest getIdccRequest(TaskDto taskDto) {
-        return gettRequest(taskDto, ProcessType.IDCC);
+        return gettRequest(taskDto);
     }
 
-    CseExportRequest getD2ccRequest(TaskDto taskDto) {
-        return gettRequest(taskDto, ProcessType.D2CC);
-    }
-
-    private CseExportRequest gettRequest(TaskDto taskDto, ProcessType processType) {
+    private CseExportRequest gettRequest(TaskDto taskDto) {
         Map<String, String> processFileUrlByType = taskDto.getProcessFiles().stream()
             .collect(HashMap::new, (m, v) -> m.put(v.getFileType(), v.getFileUrl()), HashMap::putAll);
-
         return new CseExportRequest(
             taskDto.getId().toString(),
             taskDto.getTimestamp(),
-            processType,
+            ProcessType.IDCC,
             Optional.ofNullable(processFileUrlByType.get("CGM")).orElseThrow(() -> new CseAdapterException("CGM type not found")),
             Optional.ofNullable(processFileUrlByType.get("CRAC")).orElseThrow(() -> new CseAdapterException("CRAC type not found")));
     }
