+++ cse-adapter-app/src/test/resources/application-d2cc-export.yml
@@ -1,3 +1,3 @@
 cse-adapter:
-  exchange-type: export
   process-type: D2CC
+  exchange-type: export
