+++ cse-adapter-app/src/test/resources/application-d2cc-import.yml
@@ -1,0 +1,5 @@
+cse-adapter:
+  process-type: D2CC
+  exchange-type: import
+  target-ch-fs-location: /path/to/fs/location
+  target-ch-minio-location: /path/to/minio/location
