+++ cse-adapter-app/src/test/resources/application-idcc-export.yml
@@ -1,4 +1,4 @@
 cse-adapter:
   process-type: IDCC
-  exchange-type: import
+  exchange-type: export
